LayoutTests/media/video-interruption-active-when-element-created-expected.txt

Test a <video> element crated during interruption behaves correctly.

RUN(internals.beginMediaSessionInterruption())
RUN(video = document.createElement('video'))
RUN(document.body.appendChild(video))
EVENT(canplaythrough)

RUN(video.play())
100ms timer fired...
EXPECTED (video.paused == 'true') OK
RUN(internals.endMediaSessionInterruption('MayResumePlaying'))
EVENT(playing)

END OF TEST
